refactor: added dynamic routes
