+++ src/context/product/ProductReducer.js
@@ -1,19 +1,4 @@
-export const initialFilters = {
-  selectedCategory: {},
-  sort: "",
-  priceFilterValue: "",
-  ratingFilterValue: null,
-  searchQuery: "",
-};
-
-const reducerTypes = {
-  SET_CATEGORY: "SET_CATEGORY",
-  SORT: "SORT",
-  PRICE_FILTER: "PRICE_FILTER",
-  RATING_FILTER: "RATING_FILTER",
-  SEARCH_FILTER: "PRODUCTS_SEARCH",
-  RESET_FILTERS: "RESET_FILTERS",
-};
+import { initialFilters, reducerTypes } from "helpers";
 
 export const reducerFn = (state, action) => {
   switch (action.type) {
