+++ src/helpers/sliderSettings.js
@@ -43,3 +43,37 @@ export const similarProductSlider = {
     },
   ],
 };
+
+export const featuredProductsSlider = {
+  dots: true,
+  infinite: false,
+  slidesToShow: 4,
+  arrows: true,
+  responsive: [
+    {
+      breakpoint: 1024,
+      settings: {
+        slidesToShow: 3,
+        dots: true,
+      },
+    },
+    {
+      breakpoint: 600,
+      settings: {
+        slidesToShow: 2,
+        slidesToScroll: 2,
+        initialSlide: 2,
+      },
+    },
+    {
+      breakpoint: 480,
+      settings: {
+        slidesToShow: 1,
+        slidesToScroll: 1,
+        infinite: true,
+        arrows: false,
+        initialSlide: 2,
+      },
+    },
+  ],
+};
