+++ src/pages/Cart/CartItem/CartItem.jsx
@@ -2,8 +2,8 @@ import React from "react";
 import { Link } from "react-router-dom";
 import { BsTrash } from "react-icons/bs";
 import "./CartItem.css";
-import { useUser } from "../../../context";
-import { getDiscountPercentage } from "../../../utils";
+import { useUser } from "context";
+import { getDiscountPercentage } from "utils";
 
 export const CartItem = (product) => {
   const {
