+++ src/pages/index.js
@@ -14,3 +14,4 @@ export { Orders } from "./Orders/Orders";
 export { Settings } from "./MyAccount/Settings/Settings";
 export { Checkout } from "./Cart/Checkout/Checkout";
 export { CartItems } from "./Cart/CartItems/CartItems";
+export { CartItem } from "./Cart/CartItem/CartItem";
