+++ src/services/index.js
@@ -1,4 +1,6 @@
-export * from "./data";
-export * from "./user";
+export * from "./address";
 export * from "./cart";
+export * from "./data";
+export * from "./order";
+export * from "./user";
 export * from "./wishlist";
