+++ src/utils/index.js
@@ -33,45 +33,5 @@ export const compose = (filters, ...utilFunctions) => {
     );
 };
 
-export const states = [
-  "Andaman and Nicobar Islands",
-  "Andhra Pradesh",
-  "Arunachal Pradesh",
-  "Assam",
-  "Bihar",
-  "Chandigarh",
-  "Chhattisgarh",
-  "Dadra and Nagar Haveli",
-  "Daman and Diu",
-  "Delhi",
-  "Goa",
-  "Gujarat",
-  "Haryana",
-  "Himachal Pradesh",
-  "Jammu and Kashmir",
-  "Jharkhand",
-  "Karnataka",
-  "Kerala",
-  "Lakshadweep",
-  "Madhya Pradesh",
-  "Maharashtra",
-  "Manipur",
-  "Meghalaya",
-  "Mizoram",
-  "Nagaland",
-  "Orissa",
-  "Pondicherry",
-  "Punjab",
-  "Rajasthan",
-  "Sikkim",
-  "Tamil Nadu",
-  "Tripura",
-  "Uttaranchal",
-  "Uttar Pradesh",
-  "West Bengal",
-];
-
 export * from "./filters";
 export * from "./cart";
-export * from "./sliderSettings";
-export * from "./validation";
